#593 [RENAME] Loading 폴더 구조 변경
