+++ src/components/Loading/index.js
@@ -1,2 +1,2 @@
-export { default as FetchLoading } from './FetchLoading';
-export { default as FetchLoadingOverlay } from './FetchLoadingOverlay';
+export { default as FetchLoading } from './FetchLoading/FetchLoading';
+export { default as FetchLoadingOverlay } from './FetchLoadingOverlay/FetchLoadingOverlay';
